Add source for jdk_1.2.2_006, First version for Linux.

+++ include-old/vmprofiler.h
@@ -1,15 +1,11 @@
 /*
- * @(#)vmprofiler.h	1.26 99/04/22
+ * @(#)vmprofiler.h	1.2 00/01/12
  *
- * Copyright 1997-1999 by Sun Microsystems, Inc.,
- * 901 San Antonio Road, Palo Alto, California, 94303, U.S.A.
- * All rights reserved.
+ * Copyright 1997-2000 Sun Microsystems, Inc. All Rights Reserved.
  * 
- * This software is the confidential and proprietary information
- * of Sun Microsystems, Inc. ("Confidential Information").  You
- * shall not disclose such Confidential Information and shall use
- * it only in accordance with the terms of the license agreement
- * you entered into with Sun.
+ * This software is the proprietary information of Sun Microsystems, Inc.  
+ * Use is subject to license terms.
+ * 
  */
 
 #ifndef _JAVASOFT_VMPROFILER_H_
@@ -98,8 +94,9 @@ extern jvmpi_event_info_t jvmpi_event_info[];
 #define JVMPI_EVENT_ENABLED    -2
 #define JVMPI_EVENT_DISABLED   0
 
+/* [sbb] added unsigned int cast */
 #define JVMPI_EVENT_IS_ENABLED(e) \
-  ((e) > 32 ? (jvmpi_event_info[e].flag == JVMPI_EVENT_ENABLED) : \
+  ((e) > 32 ? (jvmpi_event_info[e].flag == (unsigned int)JVMPI_EVENT_ENABLED) : \
               (jvmpi_event_flags & (1 << e)))
 
 /* constants */
